chore(alloy): cleanup network crate
